Remove incorrect early exit from ref resolver
In b3c7d24 we added support for applying the with keyword to the data
document. In the implementation, the ref resolver had to be updated to
check an additional cache. Those changes contained a bug that caused the
ref resolver to exit early and indicate that the ref was undefined if
the cache did not contain a value for the ref (partial or complete).

If the with cache does not contain a value for the ref, the value from
storage ought to be used. These changes just remove the early
exit--which is safe since the rest of the code checks if the value
returned by the with cache is non-nil (and handles the nil case
accordingly).

Fixes #1110

Signed-off-by: Torin Sandall <torinsandall@gmail.com>