Fix bug in local declaration rewriting
The rewriting was accidentally including the ref head (i.e., input or
data) in the seen stack when recursing on the right hand side of the
declaration statement. This meant that if multiple with modifiers were
applied within the term (e.g., inside a comprehension), then the 2nd
would cause the subsequent with target semantic check to fail.

The fix simply updates the rewriting step to ignore the with target in
this case (like we already do for other statements.)

Signed-off-by: Torin Sandall <torinsandall@gmail.com>