Fix with keyword invalidation
The evaluator was not invalidating the virtual document cache or
input/data documents set by the with keyword. As a result, subsequent
steps would see incorrect values. This issue goes back to
aaa5d45 but the copy propagation also required correction to ensure
that with modifiers were not thrown away on saved expressions.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>