wasm: Fix bug in composite term planning
The planner was not setting the target to the composite term before
invoking the iterator. As a result, the target was referring to the
last embedded term. For example:

planTerm([1]) would leave ltarget pointing at 1 instead of [1].

This was not noticed before because partial evaluation would remove
composite literals before the planner would see them.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>