Add livenessProbe to demo deployment
I expect many install OPA following this guide (as we did). Recent PRs
have made steps to 'productionize' this (e.g.
[#1435](https://github.com/open-policy-agent/opa/pull/1435))

We had an incident involving the controller where a stuck container was
not restarted. We would have been helped if a liveness probe was
configured. We copied the docs and this is our bad but we'd like to do
our best to make sure others don't make the same mistake.

I figured it'd be ok to use the health endpoint
[here](https://github.com/open-policy-agent/opa/blob/master/docs/content/rest-api.md#health-api)

We've made this change and it seems to be working ok for us.

Signed-off-by: Charlie Egan <charlieegan3@users.noreply.github.com>