travis: Build and push windows and darwin edge binaries
These changes update the travis build to produce windows/darwin
binaries in addition to the linux binary and push them to the S3
bucket holding the edge binaries. This lets users on windows/darwin
easily obtain the binary for tip of master. It also enables the live
docs build which is coming soon.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>